Prepare `DefaultInstantExecution` to stream the fingerprint instead of collecting it
- Extract `runToCompletion` to its own module
- Make it easier to instantiate and dispose of the `DefaultWriterContext`
