+++ subprojects/instant-execution/src/main/kotlin/org/gradle/instantexecution/serialization/Running.kt
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2020 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.gradle.instantexecution.serialization
+
+import org.gradle.instantexecution.coroutines.runToCompletion
+
+
+/**
+ * Runs the given [writeOperation] synchronously.
+ *
+ * @see runToCompletion
+ */
+internal
+fun DefaultWriteContext.runWriteOperation(writeOperation: suspend DefaultWriteContext.() -> Unit) {
+    runToCompletion {
+        writeOperation()
+    }
+}
